+++ lib/src/models/notification_schedule.dart
@@ -2,7 +2,6 @@ import 'package:awesome_notifications/src/models/model.dart';
 import 'package:awesome_notifications/src/utils/assert_utils.dart';
 import 'package:awesome_notifications/src/utils/date_utils.dart';
 import 'package:awesome_notifications/src/utils/list_utils.dart';
-import 'package:flutter/material.dart';
 
 /// Notification schedule configuration
 /// [initialDate]: (YYYY-MM-DD hh:mm:ss) The initial date that schedule should be called by first time
@@ -21,7 +20,7 @@ class NotificationSchedule extends Model {
       this.preciseSchedules});
 
   NotificationSchedule fromMap(Map<String, dynamic> dataMap) {
-    this.initialDateTime = DateUtil.parseStringToDate(
+    this.initialDateTime = DateUtils.parseStringToDate(
         AssertUtils.extractValue(dataMap, 'initialDateTime'));
     this.crontabSchedule = AssertUtils.extractValue(dataMap, 'crontabSchedule');
     this.allowWhileIdle = AssertUtils.extractValue(dataMap, 'allowWhileIdle');
@@ -32,7 +31,7 @@ class NotificationSchedule extends Model {
       preciseSchedules = List<DateTime>();
 
       for (String schedule in schedules) {
-        DateTime scheduleDate = DateUtil.parseStringToDate(schedule);
+        DateTime scheduleDate = DateUtils.parseStringToDate(schedule);
         if (schedule != null) {
           preciseSchedules.add(scheduleDate);
         }
@@ -44,7 +43,7 @@ class NotificationSchedule extends Model {
 
   Map<String, dynamic> toMap() {
     Map<String, dynamic> dataMap = {
-      'initialDateTime': DateUtil.parseDateToString(initialDateTime),
+      'initialDateTime': DateUtils.parseDateToString(initialDateTime),
       'crontabSchedule': crontabSchedule,
       'allowWhileIdle': allowWhileIdle,
       'preciseSchedules': null
@@ -54,7 +53,7 @@ class NotificationSchedule extends Model {
       List<String> schedulesMap = [];
 
       for (DateTime schedule in preciseSchedules) {
-        String scheduleDate = DateUtil.parseDateToString(schedule);
+        String scheduleDate = DateUtils.parseDateToString(schedule);
         if (schedule != null) {
           schedulesMap.add(scheduleDate);
         }